# docs/mobile-handoff/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MedicalPlab Product API",
    "description": "Fail-closed MedicalPlab API for pilot and production clients",
    "version": "1.1.0"
  },
  "paths": {
...
    "/api/v1/anatomy/command": {
      "post": {
        "summary": "Anatomy Command",
        "operationId": "anatomy_command_api_v1_anatomy_command_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnatomyCommandRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Anatomy Command Api V1 Anatomy Command Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/learn/query": {
      "post": {
        "summary": "Course Learn Query",
        "description": "Execute an evidence-grounded course learning query across medical tracks.",
        "operationId": "course_learn_query_api_v1_learn_query_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CourseLearningQueryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Course Learn Query Api V1 Learn Query Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/clinical/reason": {
      "post": {
        "summary": "Clinical Reason",
        "operationId": "clinical_reason_api_v1_clinical_reason_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClinicalReasonRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tutor/chat": {
      "post": {
        "summary": "Tutor Chat",
        "description": "Canonical Evidence-Grounded Socratic Tutor chat endpoint.",
        "operationId": "tutor_chat_api_v1_tutor_chat_post",
        "parameters": [
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TutorChatRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TutorChatResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/plab/questions/{question_id}": {
      "get": {
        "summary": "Get Plab Question",
        "operationId": "get_plab_question_api_v1_plab_questions__question_id__get",
        "parameters": [
          {
            "name": "question_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Question Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Plab Question Api V1 Plab Questions  Question Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/plab/evaluate": {
      "post": {
        "summary": "Evaluate Plab Answer",
        "operationId": "evaluate_plab_answer_api_v1_plab_evaluate_post",
        "parameters": [
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PLABEvaluationRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Evaluate Plab Answer Api V1 Plab Evaluate Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/plab/progress": {
      "get": {
        "summary": "Plab Progress",
        "operationId": "plab_progress_api_v1_plab_progress_get",
        "parameters": [
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Plab Progress Api V1 Plab Progress Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/internal/plab/reviews": {
      "get": {
        "summary": "List Plab Reviews",
        "operationId": "list_plab_reviews_api_v1_internal_plab_reviews_get",
        "parameters": [
          {
            "name": "risk_level",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Risk Level"
            }
          },
          {
            "name": "X-User-Id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-User-Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "x-user-role",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Role"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response List Plab Reviews Api V1 Internal Plab Reviews Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/internal/plab/review/status/{question_id}": {
      "get": {
        "summary": "Plab Review Status",
        "operationId": "plab_review_status_api_v1_internal_plab_review_status__question_id__get",
        "parameters": [
          {
            "name": "question_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Question Id"
            }
          },
          {
            "name": "X-User-Id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-User-Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "x-user-role",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Role"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Plab Review Status Api V1 Internal Plab Review Status  Question Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AnatomyCommandRequest": {
        "properties": {
          "query": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1000
              },
              {
                "type": "null"
              }
            ],
            "title": "Query"
          },
          "action": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "focus",
                  "highlight",
                  "isolate",
                  "ghost",
                  "reset",
                  "show",
                  "hide"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Action"
          },
          "structure_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 16,
            "title": "Structure Ids"
          },
          "opacity": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 1.0,
                "minimum": 0.05
              },
              {
                "type": "null"
              }
            ],
            "title": "Opacity"
          }
        },
        "type": "object",
        "title": "AnatomyCommandRequest"
      },
...
      "CourseLearningQueryRequest": {
        "properties": {
          "course_id": {
            "type": "string",
            "maxLength": 100,
            "minLength": 2,
            "title": "Course Id"
          },
          "query": {
            "type": "string",
            "maxLength": 4000,
            "minLength": 3,
            "title": "Query"
          },
          "intent": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "explain",
                  "question",
                  "compare",
                  "check",
                  "quiz"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Intent"
          }
        },
        "type": "object",
        "required": [
          "course_id",
          "query"
        ],
        "title": "CourseLearningQueryRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
    }
  }
}